+++ pmd-modelica/src/main/java/net/sourceforge/pmd/lang/modelica/ast/ModelicaNode.java
@@ -4,8 +4,6 @@
 
 package net.sourceforge.pmd.lang.modelica.ast;
 
-import net.sourceforge.pmd.annotation.DeprecatedUntil700;
-import net.sourceforge.pmd.lang.ast.AstVisitor;
 import net.sourceforge.pmd.lang.ast.impl.javacc.JjtreeNode;
 import net.sourceforge.pmd.lang.modelica.resolver.ModelicaScope;
 
@@ -13,17 +11,6 @@ import net.sourceforge.pmd.lang.modelica.resolver.ModelicaScope;
  * Public interface for all Modelica AST nodes.
  */
 public interface ModelicaNode extends JjtreeNode<ModelicaNode> {
-
-
-    /**
-     * @deprecated Use {@link #acceptVisitor(AstVisitor, Object)}
-     */
-    @DeprecatedUntil700
-    @Deprecated
-    default Object jjtAccept(ModelicaParserVisitor visitor, Object data) {
-        return acceptVisitor(visitor, data);
-    }
-
 
     /**
      * Returns the lexical scope this node is contained in.